requirements.txt
asgiref==3.5.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.0.4
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
filelock==3.16.0
gunicorn==20.1.0
identify==2.6.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
MarkupSafe==2.1.5
nodeenv==1.9.1
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
pre-commit==3.8.0
psycopg2-binary
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
simplejson==3.19.3
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
virtualenv==20.26.4
django-mssql-backend==2.8.1
pyodbc==5.2.0
mssql-django


